requirements.txt
numpy
matplotlib
omegaconf==2.1.1
einops==0.3.0
torch-fidelity==0.3.0
--extra-index-url https://download.pytorch.org/whl/cu116 
torch==1.12.1
torchvision==0.13.1
Pillow==9.0.1
timm==0.5.4
tqdm==4.64.0
wandb==0.12.21
torchmetrics==0.9.2
scikit-image
pytorch-lightning==1.6.5
lpips==0.1.4